update: client tests
